requirements.txt
pattern
tqdm
xlsxwriter
pyyaml
